BACKPORT: wifi: rtw89: phy: modify register setting of ENV_MNTR, PHYSTS and DIG
The ENV_MNTR(environment monitor) is the dynamic mechanism which based on
the HW of CCX(Cisco Compatible Extensions) which provide the channel
loading and noisy level indicator to debug or support the 802.11k. The
PHYSTS provide the detail PHY information per packet we received for
debugging. The DIG(dynamic initial gain) is the dynamic mechanism to
adjust the packet detect power level by received signal strength to avoid
false detection of the WiFi packet.

The address of registers used for ENV_MNTR, PHYSTS and DIG of WiFi 7 IC
are different with WiFi 6 series, so we modify the method to access the
register address in order to compatible with all WiFi 7 and 6 ICs.

Signed-off-by: Cheng-Chieh Hsieh <cj.hsieh@realtek.com>
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Kalle Valo <kvalo@kernel.org>
Link: https://lore.kernel.org/r/20230822125822.23817-7-pkshih@realtek.com
(cherry picked from commit 058b207481977c4fab7f5d979aa783cf536e21f3)

Conflicts:
   drivers/net/wireless/realtek/rtw89/phy.c
	(contextual conflicts)
   drivers/net/wireless/realtek/rtw89/rtw8851b.c
	(Unmerged path)

BUG=b:371471747
TEST=verify suite:wifi_matfunc and suite:wifi_perf on dedede with 8852BT.

Change-Id: I65d8ebad141f4fde8fb7306a55bd6766dc703286
Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5965368
Tested-by: Ruth Mekonnen <rmekonnen@google.com>
Reviewed-by: Ruth Mekonnen <rmekonnen@google.com>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>