UPSTREAM: media: amphion: Use vb2_get_buffer() instead of directly access to buffers array
Use vb2_get_buffer() instead of direct access to the vb2_queue bufs array.
This allows us to change the type of the bufs in the future.
After each call to vb2_get_buffer() we need to be sure that we get
a valid pointer so check the return value of all of them.

Signed-off-by: Benjamin Gaignard <benjamin.gaignard@collabora.com>
Reviewed-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>
CC: Ming Qian <ming.qian@nxp.com>
CC: Zhou Peng <eagle.zhou@nxp.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit e03bcb28989154de1f8998a9bdd62e9aaaed98d7)

BUG=b:348096925
TEST=CQ buildtest

Change-Id: I5c7c8a05cd93d76941a3baf9e4865170f6ea4a6d
Signed-off-by: Yunke Cao <yunkec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5712113
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>