UPSTREAM: media: mediatek: jpeg: Use vb2_get_buffer() instead of directly access to buffers array
Use vb2_get_buffer() instead of direct access to the vb2_queue bufs array.
This allows us to change the type of the bufs in the future.
After each call to vb2_get_buffer() we need to be sure that we get
a valid pointer so check the return value of all of them.

Signed-off-by: Benjamin Gaignard <benjamin.gaignard@collabora.com>
Reviewed-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>
CC: Bin Liu <bin.liu@mediatek.com>
CC: Matthias Brugger <matthias.bgg@gmail.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit a6e86185188c87465402c282fcc612b204602008)

BUG=b:348096925
TEST=None

Change-Id: Ia01c503b61c377cd3e1c580c2103f10744ed6bdc
Signed-off-by: Yunke Cao <yunkec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5649812
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>