UPSTREAM: media: dvb-core: Use vb2_get_buffer() instead of directly access to buffers array
Use vb2_get_buffer() instead of direct access to the vb2_queue bufs array.
This allows us to change the type of the bufs in the future.
After each call to vb2_get_buffer() we need to be sure that we get
a valid pointer so check the return value of all of them.

Signed-off-by: Benjamin Gaignard <benjamin.gaignard@collabora.com>
Reviewed-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit 7dc866df40127dceac9ba83ae16c0c11e7d1666f)

BUG=b:348096925
TEST=CQ buildtest

Change-Id: Ibc9bdbce988f0df492aa3678ec0c584a77b9cf57
Signed-off-by: Yunke Cao <yunkec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5715929
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>