UPSTREAM: wifi: rtw89: chan: tweak bitmap recalc ahead before MLO
Originally, we just declared two sub-entity, and according to rolling
down mechanism, we ensured that index 0 contained sub-entity as long
as there are sub-entity. So, we could use for-loop after deciding the
last index.

But, we are preparing to expand num of sub-entity for MLO. Then, there
won't be just two sub-entity. And, there might be holes between two bits
in the bitmap. So, we cannot simply do for-loop as before. Instead, we
need to follow the set bits.

Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Kalle Valo <kvalo@kernel.org>
Link: https://msgid.link/20240206030624.23382-4-pkshih@realtek.com
(cherry picked from commit ab12a3bfbf775182dc86ef8f70eab487c5deb761)

BUG=b:371471747
TEST=verify suite:wifi_matfunc and suite:wifi_perf on dedede with 8852BT.

Change-Id: Id0f0838e630f015cb8b503512e68ea989e86874b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5965521
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Ruth Mekonnen <rmekonnen@google.com>
Tested-by: Ruth Mekonnen <rmekonnen@google.com>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>