CHROMIUM: iwl7000: mvm: include link ID when releasing frames
When releasing frames from the reorder buffer, the link ID was not
included in the RX status information. This subsequently led mac80211 to
drop the frame. Change it so that the link information is set
immediately when possible so that it doesn't not need to be filled in
anymore when submitting the frame to mac80211.

Change-Id: Iec1bf5c884e371f7bc5ea2534ed9ea8d3f2c0bf6
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 5e52ed3dfa9dfa1a2833d50a7c7e882b5285d9d5