UPSTREAM: crypto: x86/aes-xts - make non-AVX implementation use new glue code
Make the non-AVX implementation of AES-XTS (xts-aes-aesni) use the new
glue code that was introduced for the AVX implementations of AES-XTS.
This reduces code size, and it improves the performance of xts-aes-aesni
due to the optimization for messages that don't span page boundaries.

This required moving the new glue functions higher up in the file and
allowing the IV encryption function to be specified by the caller.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 751fb2528c12ef64d1e863efb196cdc968b384f6)
BUG=b:349650416
TEST=Applied full stack of CLs and ran crypto self-tests on Ice Lake
Change-Id: Ib04bbb16a03a0e8e1eca8c48aa0f5678e2e4d8ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5670672
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>