UPSTREAM: drm/i915/dp_mst: Make HBLANK expansion quirk work for logical ports
The DPCD OUI of the logical port on a Dell UHBR monitor - on which the
AUX device is used to enable DSC - is all 0. To detect if the HBLANK
expansion quirk is required for this monitor use the OUI of the port's
parent instead.

Since in the above case the DPCD of both the logical port and the parent
port reports being a sink device (vs. branch device) type, read the
proper sink/branch OUI based on the DPCD device type.

This is required by a follow-up patch enabling the quirk for the above
Dell monitor.

Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Change-Id: Idcd42351d3ccfa38d7d61cf87e4bba9df9b313b7
Signed-off-by: Imre Deak <imre.deak@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240416221010.376865-11-imre.deak@intel.com
(cherry picked from commit 427c70302bbef0542a11464066447d66f97a6fe0 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)
Signed-off-by: Charlton Lin <charlton.lin@intel.com>

BUG=b:332903004, b:329139460
TEST=Verify Dell U3224KB 6k mode on Rex using UHBR, MST encoder, Big Joiner, and DSC

Change-Id: I1b0a0a9546e418f83f08fad4069ec4d311aea5b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5499990
Tested-by: Charlton Lin <charlton.lin@intel.com>
Reviewed-by: Manasi Navare <navaremanasi@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Manasi Navare <navaremanasi@chromium.org>