CHROMIUM: drm/mediatek: Fix switch secure chan fail while using secure fb
This patch will be squash to CL:5173272 and upstream together.

old_plane_state->crtc may be NULL and plane->state->crtc is not NULL
when the plane is created first time, so we changed to check
new_plane_state->crtc and new_crtc_state to fix the checking mistake.

BUG=b:343848909
TEST=emerge-geralt sys-kernel/chromeos-kernel-6_1

Change-Id: Idfbfa16a40212b0a9a373025ad183b9a61465b51
Signed-off-by: Jason-jh Lin <jason-jh.lin@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5611178
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>