CHROMIUM: arm64: dts: mt8188-geralt: Reduce secured CMA reserved memory region size
Experiments show that reserving large memory region, even being marked
as 'reusable', degrades the overall memory utilization and system
performance. The reserved memory region is only reusable for volatile
and cached data according to [1], but MTK platforms still seems to rely
on buffers that can't be migrated during runtime, e.g. DMA buffers etc.

Largely reduce the reserved memory region size for secured CMA, but keep
also the node so that we don't need to remove the associated nodes and
make it easier for devs to modify locally.

[1]: https://www.kernel.org/doc/Documentation/devicetree/bindings/reserved-memory/reserved-memory.yaml

BUG=b:339439618,b:340889609
TEST=ui.MeetCUJ.docs input latency improved

Change-Id: I77cd085a43eb3aa385dfeebcf996a4df2cf1530d
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5623718
Reviewed-by: Wei-Shun Chang <weishunc@chromium.org>
Reviewed-by: Jeff Kardatzke <jkardatzke@google.com>