CHROMIUM: media: mediatek: vcodec: Fix Capability test failure
This CL changes the following error to a warning for debugging purpose.

[MTK_V4L2][ERROR] [195]Error!! Get GET_PARAM_PICTURE_INFO Fail

This is not an error because default picture info can be used instead
for initialization. This change is needed to avoid a failure for
video.Capability Tast test. Note that this partially fixes
crrev/c/5633318.

BUG=b:344651826
TEST=tast run $DUT_IP video.Capability on ciri
TEST=tast run $DUT_IP video.ChromeStackDecoderVerification.h264_main on
ciri
TEST=tast run $DUT_IP video.ChromeStackDecoderVerification.vp9_0_group1*
on ciri
TEST=tast run $DUT_IP video.ChromeStackDecoderVerification.vp9_2_group1*
on ciri
TEST=tast run $DUT_IP video.ChromeStackDecoderVerification.av1_common on
ciri
TEST=tast run $DUT_IP
video.ChromeStackDecoderVerification.av1_10bit_common on ciri
TEST=tast run $DUT_IP
video.ChromeStackDecoderVerification.hevc_main_part_1 on ciri
TEST=tast run $DUT_IP video.PlatformDecoding.v4l2_stateless_hevc_main_10
on ciri
UPSTREAM-TASK=b:327532616

Change-Id: Iee4dee555541df4e964345d78f5c10f54287359b
Signed-off-by: Steve Cho <stevecho@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5668552
Tested-by: Steve Cho <stevecho@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>