CHROMIUM: iwl7000: mld: improve hcmd functions
In most cases, sending a command to the firmware is
just using "&cmd" as the command, and sizeof(cmd)
for the size. Make the hcmd functions macros that
need not be invoved with the size, so that in such
cases we can simply write

  iwl_mld_send_cmd_pdu(mld, ..., &cmd);

but also continue writing

  iwl_mld_send_cmd_pdu(mld, ..., ptr, size);

To do this, pull the code into the header and also
remove the C file, but that wasn't really needed
in the first place since it's just trivial code.

Change-Id: I8b44ee93496e14d7ed3e674653f861c884eb71da
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: edfb63e0a88fcefb4907d175f09d6550e20d2ac0