CHROMIUM: iwl7000: mvm: don't wait for tx queues if firmware is dead
There is a WARNING in iwl_trans_wait_tx_queues_empty() which we hit it's
become a WARNING.
The problem is that we can't expect anything from the firmware after
it's declared dead.
Don't call iwl_trans_wait_tx_queues_empty() in this case. While it could
be a good idea to stop the flow earlier, the flush functions do some
maintenance work that is not related to the firmware, so keep that part
of the code running even when the firmware is not running.

Change-Id: I44a739fbd4ffcc46b83844dd1c7b2eb0c7b270f6
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 850e413d9b25201ecaea4567b463ef855ac0c34f