CHROMIUM: uapi: drm: DRM_MTK_GEM_CREATE_RESTRICTED as bitwise flag
DRM_MTK_GEM_CREATE_RESTRICTED is not a DRM IOCTL command; it's a flag
of DRM_MTK_GEM_CREATE. Assigning it a sequential code in addition to
other DRM IOCTL commands is wrong.

Rename the flag with clearer name and give it a bitwise value.

Note that this patch is NOT upstreamable because the idea of allocating
graphics related buffer via DRM GEM create ioctl calls is a
ChromeOS-on-MTK tweak and was rejected by upstream, so the base uAPI
header files doesn't exist in upstream.
We have to carry these changes as CHROMIUM.

BUG=b:352229429
TEST=emerge-geralt chromeos-kernel-6_1

Cq-Depend: chromium:5689125
Change-Id: I6320f2f9709a63c34c952ffbbc963708b01793cd
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5691089
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>