CHROMIUM: iwl7000: mld: reset deflink in fw upon disassoc
The deflink is alive while the vif is. For non-MLO, this will be the only
link in fw. For MLO, deflink will not exist in fw.

After a non-MLO connection, we need to clear deflink's state in fw.

For MLO connections, deflink will be removed/added from the fw at the
beginning/end of the connection, respectively. This happens in
drv_change_vif_links, which is not called in non-MLO.

So remove and re-add deflink in fw, in drv_unassign_chanctx - after it is
deactivated, to clear its state.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I8390c624819761d29569ee6c873d89a30d7d5c4a
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: de28c33f4e4417a58c09b45aeab6a9b95775f7da