CHROMIUM: arm64: dts: mt8188-geralt: Remove bogus backlight_lcd1
backlight_lcd1 was described as the backlight supply of Geralt Proto's
eDP panel.
However that panel is an OLED panel and doesn't require a backlight
driver. The OLED driver should control the panel brightness directly.

Remove the faulty backlight_lcd1 and its connections.
After which, since there's only one backlight_lcd0, rename it to
backlight_lcd.

Furthermore, the backlight power supply is actually from the source
system power rail but not 5.0V. Correct that in the node and remove the
spurious pp5000_bl regulator.

BUG=b:311139883,b:317303448
TEST=emerge kernel and both Geralt Proto and Ciri boot
UPSTREAM-TASK=b:244281402

Change-Id: I30fa2c3a6af56f5acaf07927a63d2ef89a1c6306
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5759062
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>