CHROMIUM: iwl7000: mld: avoid kernel panic in restart
p2p device is not supported yet, so we don't add the driver part of it.
But in iwl_mld_restart_cleanup we iterate all the interface to clean
them up, this will cause a kernel panic for the p2p interface, since the
driver part of it is uninitialized.

Fix this by not doing anything in the cleanup iterator for interfaces
that are not bss stations. Will be reverted when more interface types
will be supported.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I248e973cff166663c944284ec6e8e78a8a9fbbf1
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: c05cdcd7165cc2914efc60d7cd18e37c3b9757d3