CHROMIUM: thermal: mediatek: Add virtual temp driver (v2)
Add virtual temp driver to create a virtual sensor for the thermal
aggregation zone (soc_max) for device thermal controls.

This driver is a refactored version based on previous MTK driver.
MTK attempted to upstream this but got rejected, and they are working on
the alternative solution. This is the tentative solution before that
becomes available.

BUG=b:236331727
UPSTREAM-TASK=b:236331727
TEST=verify thermal zone on MT8188 and MT8192
     grep -R '' /sys/class/thermal/thermal_zone*/{type,temp}

Change-Id: I363a74a4bbc00dea6218e3e0955edc179b28a0f4
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5785499
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>