CHROMIUM: iwl7000: mld: add the right link_sta in reconfig
When a STA is added, it has only the deflink link_sta, which has the
link id of the association link.
After assoc, mac80211 will allocate the other link_sta, and when we enter
EMLSR, the driver will add the other link_sta to the FW.

So, when the STA is added we can add only the deflink link_sta to the FW.

But that's true only for the normal flow.

In restart while in EMLSR, mac80211 will switch to one active link -
to ffs(active_links), which doesn't necessarily have the same link id as
the deflink link_sta has. This link will be the only one added to the
FW (at this stage. At the end of the restart, mac80211 will switch back
to EMLSR).

In that case, we will add to the FW a STA with a link id that doesn't
exist in the FW yet.

Fix this by using for_each_sta_active_link when adding a STA, since this
checks if the link_sta's link ID is in active_links. This will work for
both the normal and restart flows.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I4068d9a71e9607b3a68126ec021ed60c139aab8a
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: d4e6e62b19f739b317a6f18fe890a6bc230bcd79