CHROMIUM: arm64: dts: mediatek: mt8188: Remove maxim,dsm_param_name property
The `maxim,dsm_param_name` property is for ACPI, not DT. It's possible
that ODM integrated the configuration from existing projects with ACPI,
but didn't clean this up.

The max98390 driver has a default name `dsm_param.bin` for the DSM
parameter files, and that aligns to what MT8188 has in /lib/firmware/,
so this change should have no practical impact.

Remove this property since it's no-op and shouldn't be used on any DT.

BUG=b:376376802
UPSTREAM-TASK=b:244281402
TEST=Ciri SKU1 (Maxim smartamp) speaker works normally

Change-Id: Icbb82cf66e19cb2444a53d769f146a19ec561b4e
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5974803
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>